+++ src/app/pages/about/about.page.ts
@@ -2,7 +2,6 @@ import { Component, Inject, LOCALE_ID, OnInit } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 import { ActivatedRoute } from '@angular/router';
 import { catchError, map, Observable, of, switchMap } from 'rxjs';
-import { marked } from 'marked';
 
 import { MarkdownContentService } from '@services/markdown-content.service';
 
@@ -33,7 +32,9 @@ export class AboutPage implements OnInit {
   getMdContent(fileID: string): Observable<SafeHtml> {
     return this.mdContentService.getMdContent(fileID).pipe(
       map((res: any) => {
-        return this.sanitizer.bypassSecurityTrustHtml(marked(res.content));
+        return this.sanitizer.bypassSecurityTrustHtml(
+          this.mdContentService.getParsedMd(res.content)
+        );
       }),
       catchError((e: any) => {
         console.error('Error loading markdown content', e);