+++ src/app/pages/home/home.page.ts
@@ -2,7 +2,6 @@ import { Component, Inject, LOCALE_ID, OnInit } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 import { Router } from '@angular/router';
 import { catchError, map, Observable, of } from 'rxjs';
-import { marked } from 'marked';
 
 import { config } from '@config';
 import { MarkdownContentService } from '@services/markdown-content.service';
@@ -67,7 +66,9 @@ export class HomePage implements OnInit {
   private getMdContent(fileID: string): Observable<SafeHtml> {
     return this.mdContentService.getMdContent(fileID).pipe(
       map((res: any) => {
-        return this.sanitizer.bypassSecurityTrustHtml(marked(res.content));
+        return this.sanitizer.bypassSecurityTrustHtml(
+          this.mdContentService.getParsedMd(res.content)
+        );
       }),
       catchError((e) => {
         console.error(e);