+++ src/app/services/markdown-content.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { catchError, map, Observable, of } from 'rxjs';
+import { marked } from 'marked';
 
 import { config } from '@config';
 
@@ -41,6 +42,10 @@ export class MarkdownContentService {
         return of({});
       })
     );
+  }
+
+  getParsedMd(md: string): string {
+    return marked.parse(md) as string;
   }
 
   /**